pyproject.toml
[tool.poetry]
name = "dronesim-python"
version = "0.1.0"
description = ""
authors = ["SKYnSPACE <skynspace@kaist.ac.kr>"]
readme = "README.md"
packages = [{include = "dronesim_python"}]

[tool.poetry.dependencies]
python = "^3.11"
pyserial = "^3.5"
control = "^0.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
